+++ pytsite/auth_storage_odm/__init__.py
@@ -25,11 +25,11 @@ def _init():
     # 'Users' admin sidebar menu
     url = router.ep_path('pytsite.odm_ui@browse', {'model': 'user'})
     admin.sidebar.add_menu('auth', 'users', 'pytsite.auth_storage_odm@users', url, 'fa fa-users', weight=10,
-                           permissions='pytsite.odm_perm.view.user')
+                           permissions='pytsite.odm_auth.view.user')
 
     # 'Roles' admin sidebar menu
     url = router.ep_path('pytsite.odm_ui@browse', {'model': 'role'})
     admin.sidebar.add_menu('auth', 'roles', 'pytsite.auth_storage_odm@roles', url, 'fa fa-key', weight=20,
-                           permissions='pytsite.odm_perm.view.role')
+                           permissions='pytsite.odm_auth.view.role')
 
 _init()