+++ pytsite/content/res/assets/js/widget/entity-select.js
@@ -1,3 +1,0 @@
-$(window).on('pytsite.widget.init:pytsite.content._widget.EntitySelect', function (e, widget) {
-    $(window).trigger('pytsite.widget.init:pytsite.widget._select.Select2', [widget]);
-});