+++ pytsite/odm_auth/__init__.py
@@ -10,11 +10,11 @@ __license__ = 'MIT'
 
 
 def _init():
-    from pytsite import events, http_api
+    from pytsite import events
     from . import _eh
 
     # Event listeners
-    events.listen('pytsite.odm.register_model', _eh.odm_register_model)
+    events.listen('pytsite.odm.register', _eh.odm_register_model)
     events.listen('pytsite.odm.entity.pre_save', _eh.odm_entity_pre_save)
     events.listen('pytsite.odm.entity.pre_delete', _eh.odm_entity_pre_delete)